EconomyGCDModel_A1API

EconomyGCDModel_A1API

Index

EconomyGCDModel_A1 Components

Index

Ai4EComponentLib.EconomyGCDModel_A1.Capitalflow

Ai4EComponentLib.EconomyGCDModel_A1.Firm

Ai4EComponentLib.EconomyGCDModel_A1.Household

EconomyGCDModel_A1 Components

Ai4EComponentLib.EconomyGCDModel_A1.Capitalflow — Method

The port through which money flows.

States:

p(t): Price of good

w(t): Wage(per hour)

l(t): Labor

c(t): Consumption

𝛌₁(t),𝛌₂(t),𝛌₃(t):Parameters that result from the constraint force for the change in consumption and varies over time.

Parameters:

α,β: Parameters in a simple Cobb-Douglas production function.

K₀: Initial value of Capital.

Ai4EComponentLib.EconomyGCDModel_A1.Firm — Method

Component: an agent–-firm.

Function of this component:

\[DP=\hat{dp}(K)\\ firm.m.p\times firm.m.C-firm.m.w\times firm.m.L-firm.m.M=0\\ \beta(firm.m.L)^\alpha(K)^{1-\alpha}-firm.m.C-(K)'-(S)'-DP=0\\ (K)'=\mu^H_Kp\beta(1-\alpha)(firm.m.L)^\alpha(K)^{-\alpha}-\lambda_3\\ (firm.m.M)'=-\lambda_2\ (S)'=\mu^F_S(\hat{S}-S)-\lambda_3\ \ (firm.m.p)'=\mu^F_p\beta(K)^{1-\alpha}(firm.m.L)^\alpha-\lambda_1(firm.m.C)+\lambda_2(firm.m.C)\\ (firm.m.w)'=\mu^F_w(firm.m.L)+\lambda_1(firm.m.L)-\lambda_2(firm.m.L)\\ (firm.m.L)'=\mu^H_L(\hat{L}-firm.m.L)+\mu^F_L(\alpha\beta(K)^{1-\alpha}(firm.m.L)^{\alpha-1}-firm.m.w)\\ +\lambda_1(firm.m.w)-\lambda_2(firm.m.w)+\lambda_3\alpha\beta(firm.m.L)^{\alpha-1}(K)^{1-\alpha}\]

Parameters:

α,β: Parameters in a simple Cobb-Douglas production function.

dp̂: depreciation.

μ₂: The parameter of Capital.

μ₃: The parameter of Labor and Household.

μ₄: The parameter of Labor and Firm.

μ₇: The parameter of Price.

μ₈: The parameter of storage.

μ₉: The parameter of Wage.

Ŝ: The planned inventory of storage.

l̂: The wish of working time.

Connectors:

m: money.

Ai4EComponentLib.EconomyGCDModel_A1.Household — Method

Component: an agent–-Household.

Function of this component:

\[household.m.w\times household.m.L-household.m.p\times household.m.C-household.m.M\\ (household.m.M)'=2\mu^H_{M^H}(\hat{M}^H-household.m.M)-\lambda_2\\ (household.m.C)'=\mu^H_C\gamma (household.m.C)^{\gamma-1}-\lambda_1(household.m.p)+\lambda_2(household.m.p)-\lambda_3\]

States:

Mʰ(t): Money stock household.

mʰ(t): Derivative of Mʰ(t).

Parameters:

α,β: Parameters in a simple Cobb-Douglas production function.

dp̂: depreciation.

μ₁: The parameter of Consumption.

μ₆: The parameter of money stock household.

K₀: Initial value of Capital.

M̂ʰ: The amount of cash that Households aim to keep.

γ: The parameter of Consumption and Household.

Connectors:

m: money.