ElectrochemistryAPI

ElectrochemistryAPI

Index

Electrochemistry Components

Index

Ai4EComponentLib.Electrochemistry.Ground

Ai4EComponentLib.Electrochemistry.OnePort

Ai4EComponentLib.Electrochemistry.PEMElectrolyzer

Ai4EComponentLib.Electrochemistry.PhotovoltaicCell

Ai4EComponentLib.Electrochemistry.Pin

Electrochemistry Components

Ai4EComponentLib.Electrochemistry.Ground — Method

Ground(; name)

Ground node with the potential of zero and connector g. Every circuit must have one ground node.

From ModelingToolkitStandardLibrary.jl

Connectors:

g

Ai4EComponentLib.Electrochemistry.OnePort — Method

OnePort(; name)

Component with two electrical pins p and n and current i flows from p to n.

From ModelingToolkitStandardLibrary.jl

States:

v(t): [V] The voltage across component p.v - n.v

i(t): [A] The current passing through positive pin

Parameters:

v_start: [V] Initial voltage across the component

i_start: [A] Initial current through the component

Connectors:

p Positive pin

n Negative pin

Ai4EComponentLib.Electrochemistry.PEMElectrolyzer — Method

PEMElectrolyzer(; name, τ, I_0, n, u_0, R_sol)

Component: PEMElectrolyzer(Proton Exchange Membranes Electrolyzer)

I-U in Electrolyzer:

\[U_{cell}=U_{0}-\tau \times \sinh \left(\frac{I}{I_{0}(T)}\right)+I \times R_{\text{sol}}(T)\]

Molar Yield of Hydrogen and Oxygen, Molar consumption of water:

\[\begin{aligned} \dot{n}_{H_{2} O} &=\frac{n_{c} I_{e l}}{2 \cdot F} \\ \dot{n}_{H_{2}} &=\frac{n_{c} I_{e l}}{2 \cdot F} \cdot \eta_{f} \\ \dot{n}_{O_{2}} &=\frac{n_{c} I_{e l}}{2 \cdot F} \cdot \eta_{f} \end{aligned}\]

Mass yield of hydrogen[1]:

\[\frac{\mathrm{d}^{2} \dot{m}}{\mathrm{~d} t^{2}}=-2 \zeta \cdot \omega_{n} \frac{\mathrm{d} \dot{m}}{\mathrm{~d} t}-\omega_{n}{ }^{2} \dot{m}+\omega_{n}{ }^{2} M_{H_{2}} n_{H_{2}}\]

States:

v(t): [V] Voltage across the Electrolyzer

i(t): [A] Current through the Electrolyzer

n_H_2(t): [mol/s] Molar Yield of Hydrogen

η_f(t): Faraday efficiency, here $η_f = 1$

m_H_2(t): [g/s] Mass yield of hydrogen

∂_m_H_2: Derivative of hydrogen mass yield

Parameters:

τ: Tafel slope

I_0: Exchange Current Density

n: Number of electrolyzers in series

u_0: Equilibrium potential

R_sol: Solution resistance

Inside Variables:

ωₙ: Natural frequency of Hydrogen production [1]

ζ: Damping ratio of Hydrogen production [1]

molarMassH2: [g/mol] Molar mass of hydrogen

Connectors:

p Positive pin

n Negative pin

Reference:

[1]Espinosa-López M, Darras C, Poggi P, et al. Modelling and experimental validation of a 46 kw pem high pressure water electrolyzer[J]. Renewable energy, 2018, 119:160-173.

Ai4EComponentLib.Electrochemistry.PhotovoltaicCell — Method

PhotovoltaicCell(; name, I_ph, I_0, a, R_s, R_sh)

Component: PhotovoltaicCell with 5 Paramenters

States:

v(t): [V] Voltage across the Cell

i(t): [A] Current through the Cell

Paramenters:

I_ph: Photocurrent

I_0: Reverse saturation current

a: Correction factor

R_s: Series resistance

R_sh: Ohmic resistance

Connectors:

p Positive pin

n Negative pin

Reference: De Soto W, Klein S A, Beckman W A. Improvement and validation of a model for photovoltaic array performance[J]. Solar energy, 2006, 80(1):78-88.

Ai4EComponentLib.Electrochemistry.Pin — Method

Pin(; name)

A pin in an analog circuit.

From ModelingToolkitStandardLibrary.jl

States:

v(t): [V] The voltage at this pin

i(t): [A] The current passing through this pin