ElectrochemistryAPI

ElectrochemistryAPI

Index

Electrochemistry Components

Index

Ai4EComponentLib.Electrochemistry.Ground

Ai4EComponentLib.Electrochemistry.OnePort

Ai4EComponentLib.Electrochemistry.PhotovoltaicCell

Ai4EComponentLib.Electrochemistry.Pin

Electrochemistry Components

Ai4EComponentLib.Electrochemistry.Pin — Function

Pin(; name)

A pin in an analog circuit.

From ModelingToolkitStandardLibrary.jl

States:

v(t): [V] The voltage at this pin

i(t): [A] The current passing through this pin

Ai4EComponentLib.Electrochemistry.OnePort — Function

OnePort(; name)

Component with two electrical pins p and n and current i flows from p to n.

From ModelingToolkitStandardLibrary.jl

States:

v(t): [V] The voltage across component p.v - n.v

i(t): [A] The current passing through positive pin

Parameters:

v_start: [V] Initial voltage across the component

i_start: [A] Initial current through the component

Connectors:

p Positive pin

n Negative pin

Ai4EComponentLib.Electrochemistry.Ground — Function

Ground(; name)

Ground node with the potential of zero and connector g. Every circuit must have one ground node.

From ModelingToolkitStandardLibrary.jl

Connectors:

g

Missing docstring.

Missing docstring for Electrolyzer. Check Documenter's build log for details.

Ai4EComponentLib.Electrochemistry.PhotovoltaicCell — Function

PhotovoltaicCell(; name, I_ph, I_0, a, R_s, R_sh)

Component: PhotovoltaicCell with 5 Paramenters

States:

v(t): [V] Voltage across the Cell

i(t): [A] Current through the Cell

Paramenters:

I_ph: Photocurrent

I_0: Reverse saturation current

a: Correction factor

R_s: Series resistance

R_sh: Ohmic resistance

Connectors:

p Positive pin

n Negative pin

Reference: De Soto W, Klein S A, Beckman W A. Improvement and validation of a model for photovoltaic array performance[J]. Solar energy, 2006, 80(1):78-88.