IncompressiblePipeAPI

IncompressiblePipeAPI

Index

IncompressiblePipe Components

Index

Ai4EComponentLib.IncompressiblePipe.CentrifugalPump

Ai4EComponentLib.IncompressiblePipe.ElbowPipe

Ai4EComponentLib.IncompressiblePipe.PipeNode

Ai4EComponentLib.IncompressiblePipe.SimplePipe

Ai4EComponentLib.IncompressiblePipe.Sink_P

Ai4EComponentLib.IncompressiblePipe.Source_P

Ai4EComponentLib.IncompressiblePipe._NodeEnergy

Ai4EComponentLib.IncompressiblePipe._h_f

Ai4EComponentLib.IncompressiblePipe._h_m

IncompressiblePipe Components

Ai4EComponentLib.IncompressiblePipe.CentrifugalPump — Method

CentrifugalPump(; name, D, ω, c_0, c_1, ρ)

Component: CentrifugalPump

Ideal H-Q Characteristic curves of Centrifugal Pumps:

\[H_t=\frac{(r\omega)^2}{g}-\frac{\omega \cot\beta }{2\pi bg}Q=c_0\omega^2-c_1\omega Q=a_0-a_1Q\]

Parameters:

D: [m] Diameter of pipe

ω: [r/min] rotary speed

c_0: parameter in H-Q Characteristic curves

c_1: parameter in H-Q Characteristic curves

Connectors:

in: Inlet of pump

out: Outlet of pump

Ai4EComponentLib.IncompressiblePipe.ElbowPipe — Method

ElbowPipe(; name, D, K, ρ, zin, zout)

Component: ElbowPipe(pipe with fixed local resistance loss coefficient f)

Energy conservation equation in the form of Bernoulli Equation between two ports:

\[\frac{p_{in}}{\rho g} +\frac{8q_{in}^{2}}{\pi^2D^4g} + z_{in}= \frac{p_{out}}{\rho g} +\frac{8q_{out}^{2}}{\pi^2D^4g} + z_{out}+h_f+h_m\]

Parameters:

D: [m] Diameter of pipe

K: Local resistance loss coefficient

Connectors:

in: Inlet of pipe

out: Outlet of pipe

Arguments:

zin: [m] The height of inlet port

zout: [m] The height of outlet port

ρ: [`m³/kg] The density of fluid passing the port

Ai4EComponentLib.IncompressiblePipe.PipeNode — Method

PipeNode(; name, z)

A pipe port(inlet or outlet) in an pipe network.

States:

p(t): [Pa] The pressure at this port

q(t): [m³/s] The volume flow passing through this port

Parameters:

z: [m] The hight of port, expressing potential energy

Ai4EComponentLib.IncompressiblePipe.SimplePipe — Method

SimplePipe(; name, L, D, f, ρ, zin, zout, K_inside)

Component: SimplePipe(pipe with fixed friction factor f)

Energy conservation equation in the form of Bernoulli Equation between two ports:

\[\frac{p_{in}}{\rho g} +\frac{8q_{in}^{2}}{\pi^2D^4g} + z_{in}= \frac{p_{out}}{\rho g} +\frac{8q_{out}^{2}}{\pi^2D^4g} + z_{out}+h_f+h_m\]

Parameters:

D: [m] Diameter of pipe

L: [m] Length of pipe

f: Friction factor

K_inside: Coefficient of local resistance loss inside the pipe

Connectors:

in: Inlet of pipe

out: Outlet of pipe

Arguments:

zin: [m] The height of inlet port

zout: [m] The height of outlet port

ρ: [`m³/kg] The density of fluid passing the port

Ai4EComponentLib.IncompressiblePipe.Sink_P — Method

Sink_P(; name, p)

Component: Sink_P

Sink_P can be defined as a source(where fluids are from) or sink(where fluid are going to).

Connectors:

port: port of sink

Arguments:

p: [Pa] The pressure of sink, default: 101325 (standard atmospheric pressure)

Ai4EComponentLib.IncompressiblePipe.Source_P — Method

Source_P(; name, D, z, ρ, p, K_inlet)

Component: Source_P(source with inlet pressure losses)

Parameters:

D: [m] Diameter of pipe

K_inlet: Local resistance loss coefficient of Inlet port, default: 0.5

Connectors:

port: port of source

Arguments:

z: [m] The height of source

ρ: [m³/kg] The density of fluid

p: [Pa] The pressure of source, default: 101325 (standard atmospheric pressure)

Ai4EComponentLib.IncompressiblePipe._NodeEnergy — Method

_NodeEnergy(node, D, ρ) -> Any

To get the energy at the port.

The governing equation of incompressible pipe network is Bernoulli Equation:

\[\frac{p}{\rho g} +\frac{v^{2}}{2g} + z=\mathrm{constant}\]

In volume flow form:

\[\frac{p}{\rho g} +\frac{8q^{2}}{\pi^2D^4g} + z=\mathrm{constant}\]

D: [m] Diameter of pipe

ρ: [`m³/kg] The density of fluid passing the port

Ai4EComponentLib.IncompressiblePipe._h_f — Method

_h_f(node, f, L, D) -> Any

To get the loss of resistance along the pipe(between two ports).

In volume flow form:

\[h_f = f\frac{L}{D} \frac{8q^{2}}{\pi^2D^4g}\]

Ai4EComponentLib.IncompressiblePipe._h_m — Method

_h_m(node, K, D) -> Any

To get the local resistance loss the components.

In volume flow form:

\[h_m = K \frac{8q^{2}}{\pi^2D^4g}\]