CompressedAirSystemAPI

CompressedAirSystemAPI

Index

CompressedAirSystem Components

Index

Ai4EComponentLib.CompressedAirSystem.AirStroageTank

Ai4EComponentLib.CompressedAirSystem.ConstantValve

Ai4EComponentLib.CompressedAirSystem.FlowPort

Ai4EComponentLib.CompressedAirSystem.Purifier

Ai4EComponentLib.CompressedAirSystem.SISOComponent

Ai4EComponentLib.CompressedAirSystem.SimplifiedPipe

Ai4EComponentLib.CompressedAirSystem.Source

Ai4EComponentLib.CompressedAirSystem.StraightPipe

Ai4EComponentLib.CompressedAirSystem.TransitionPipe

Ai4EComponentLib.CompressedAirSystem.VarySpeedCompressor

Ai4EComponentLib.CompressedAirSystem.pipeFriction_turbulent

Ai4EComponentLib.CompressedAirSystem.pipeRe

Ai4EComponentLib.CompressedAirSystem.pipeVelocity

Ai4EComponentLib.CompressedAirSystem.stateEquation

CompressedAirSystem Components

Ai4EComponentLib.CompressedAirSystem.AirStroageTank — Method

AirStroageTank(; name, V, T, p0)

Air tank.

Valve with constant resistance.

Arguments:

V: [m³] Volume, defaults: 20

T: [K] Temperature, defaults: 300

p0: [Pa] Initial pressure, defaults: 1e5

Connectors:

inlet inlet of components

outlet outlet of components

Ai4EComponentLib.CompressedAirSystem.ConstantValve — Method

ConstantValve(; name, R)

Valve with constant resistance.

Connectors:

inlet inlet of components

outlet outlet of components

Ai4EComponentLib.CompressedAirSystem.FlowPort — Method

FlowPort(; name)

Port(inlet or outlet) in CompressedAirSystem.

States:

p(t): [Pa] The pressure at this port

qm(t): [kg/s] The mass flow rate passing through this port

T(t): [K] The mass flow rate passing through this port

ρ(t): [kg/m³] The density passing at this port

μ(t): [Pa⋅s] The absolute viscosity at this port

qv(t): [m³/s] The volume flow rate passing through this port

Ai4EComponentLib.CompressedAirSystem.Purifier — Method

Purifier(; name, c, Δp, T)

General purifier.

Arguments:

c: Purification coefficient, percentage of residual mass flow after purification to inlet mass flow, defaults: 1.0

Δp: [Pa] Absolute pressure drop across the component

T: [K] Outlet temperature after cooling

Connectors:

inlet inlet of components

outlet outlet of components

Ai4EComponentLib.CompressedAirSystem.SISOComponent — Method

SISOComponent(; name)

The Component with two ports inlet and outlet and mass flow qm flows from inlet to outlet. It plays the same role as Oneport in the circuit system.

States:

qm(t): [kg/s] The mass flow rate passing through this component

Δp(t): [Pa] The pressure difference at a component

ρ_mean(t): [kg/m³] The density at a component

μ_mean(t): [Pa⋅s] The absolute viscosity at a component

qv_mean(t): [m³/s] The volume flow rate passing through this component

Connectors:

inlet inlet of components

outlet outlet of components

Ai4EComponentLib.CompressedAirSystem.SimplifiedPipe — Method

SimplifiedPipe(; name, R0)

Simplified pipe model

Arguments:

R0: [kg/m^7] Resistance coefficient of pipeline

Connectors:

inlet inlet of components

outlet outlet of components

Ai4EComponentLib.CompressedAirSystem.Source — Method

Source(; name, boundary)

Source is a general source, which can generate pressure boundary("p"), temperature boundary("T") or mass flow boundary("qm") with different input parameters.

Valve with constant resistance.

Arguments:

boundary: Dict with boundaries

Connectors:

source: port of source

inletBoundary = Dict(
    "p" => 1.0e5(1 + 0.001sin(t)),
    "T" => 300,
)

Ai4EComponentLib.CompressedAirSystem.StraightPipe — Method

StraightPipe(; name, D, L, ϵ)

Straight round pipe and height difference is not considered. Friction coefficient is obtained from Modi diagram.

Arguments:

D: [m] Pipe diameter, defaults: 1.0

L: [m] Pipe length, defaults: 1.0

ϵ: [m] Absolute roughness of pipeline, defaults: 0.05

Connectors:

inlet inlet of components

outlet outlet of components

Ai4EComponentLib.CompressedAirSystem.TransitionPipe — Method

TransitionPipe(; name, n, f, D, L, T, p0, qm0)

Simplified pipe model

Arguments:

f: [kg/m⁷] Resistance coefficient of pipeline

n: Number of pipe discrete node

D: [m] Pipe diameter, defaults: 1.0

L: [m] Pipe length, defaults: 1.0

T: [K] Ambient temperature, defaults: 300

p0: [Pa] Initial pressure of each node

qm0: [kg/(m²⋅s)] Initial specific momentum of each node

Connectors:

inlet inlet of components

outlet outlet of components

Ai4EComponentLib.CompressedAirSystem.VarySpeedCompressor — Method

VarySpeedCompressor(; name, n0, h_polCoff, etaCoff, surgeCoff, chokeCoff)

Variable speed compressor.

\[h_polCoff = a1 * inlet.qv^2 + a2 * inlet.qv * n / n0 + a3 * (n / n0)^3) * g\]

\[etaCoff = b1 * inlet.qv^2 * (n0 / n)^2 + b2 * inlet.qv * n0 / n + b3\]

\[surgeCoff = (c1 * n * n + c2 * n + c3) / 3600\]

\[chokeCoff = (d1 * n * n + d2 * n + d3) / 3600\]

States:

qm(t): [kg/s] The mass flow rate

n(t): [rpm] speed

ϵ(t): pressure ratio

h_pol(t): [J] Energy head

h_tot(t): [J] Total Work

eta_pol(t): Surge limit coefficient

qv_surge(t): [kg/s] Surge flow rate

qv_choke(t): [kg/s] choke flow rate

σ(t)

Arguments:

n0: [rpm] Rated speed

h_polCoff: Work coefficient

etaCoff: Efficiency coefficient

surgeCoff: Surge limit coefficient

chokeCoff: Choke limit coefficient

Connectors:

inlet inlet of components

outlet outlet of components

Ai4EComponentLib.CompressedAirSystem.pipeFriction_turbulent — Method

pipeFriction_turbulent(f, Re, ϵ, D) -> Any

Fitting Formula for Friction Coefficient of Turbulent Pipe

Ai4EComponentLib.CompressedAirSystem.pipeRe — Method

pipeRe(ρ, u, D, μ) -> Any

Reynolds number

Ai4EComponentLib.CompressedAirSystem.pipeVelocity — Method

pipeVelocity(Δp, ρ, L, D, f) -> Any

Flow rate in pipe

Ai4EComponentLib.CompressedAirSystem.stateEquation — Method

stateEquation(p, ρ, T) -> Any

State Equation of Dry Air